Minor fixes in shared.h, figured using doubles might be best in there. Also removed an accidental #if #else its else situation, causing enities to never think.... (That's bad for business.)

+++ inc/shared/shared.h
@@ -212,8 +212,8 @@ constexpr int32_t WORLD_SIZE = (MAX_WORLD_COORD - MIN_WORLD_COORD);
 #if TICK_RATE_STYLE == 1
 static constexpr uint32_t   BASE_FRAMERATE =  10;
 static constexpr double     BASE_FRAMETIME = 100;
-static constexpr double     BASE_1_FRAMETIME = 0.01f;   // 1/BASE_FRAMETIME
-static constexpr double     BASE_FRAMETIME_1000 =0.1f;    // BASE_FRAMETIME/1000
+static constexpr double     BASE_1_FRAMETIME = 0.01;   // 1/BASE_FRAMETIME
+static constexpr double     BASE_FRAMETIME_1000 =0.1;    // BASE_FRAMETIME/1000
 #endif
 
 //-----------------
@@ -221,9 +221,9 @@ static constexpr double     BASE_FRAMETIME_1000 =0.1f;    // BASE_FRAMETIME/1000
 //-----------------
 #if TICK_RATE_STYLE == 2
 static constexpr uint32_t   BASE_FRAMERATE = 20; //10
-static constexpr double     BASE_FRAMETIME = 50.0f; //100
-static constexpr double     BASE_1_FRAMETIME = 0.02f; //0.01f   // 1/BASE_FRAMETIME
-static constexpr double     BASE_FRAMETIME_1000 = 0.05f; //0.1f    // BASE_FRAMETIME/1000
+static constexpr double     BASE_FRAMETIME = 50.0; //100
+static constexpr double     BASE_1_FRAMETIME = 0.02; //0.01f   // 1/BASE_FRAMETIME
+static constexpr double     BASE_FRAMETIME_1000 = 0.05; //0.1f    // BASE_FRAMETIME/1000
 #endif
 //-----------------
 //N&C 30hz tick
@@ -239,18 +239,18 @@ static constexpr double     BASE_FRAMETIME_1000 = 0.03333333333; //0.1f    // BA
 //-----------------
 #if TICK_RATE_STYLE == 4
 static constexpr uint32_t   BASE_FRAMERATE = 40; //10
-static constexpr float      BASE_FRAMETIME = 25.0f; //100
-static constexpr float      BASE_1_FRAMETIME = 0.04f; //0.01f   // 1/BASE_FRAMETIME
-static constexpr float      BASE_FRAMETIME_1000 = 0.025f; //0.1f    // BASE_FRAMETIME/1000
+static constexpr double     BASE_FRAMETIME = 25.0; //100
+static constexpr double     BASE_1_FRAMETIME = 0.04; //0.01f   // 1/BASE_FRAMETIME
+static constexpr double     BASE_FRAMETIME_1000 = 0.025; //0.1f    // BASE_FRAMETIME/1000
 #endif
 //-----------------
 //N&C 60hz tick
 //-----------------
 #if TICK_RATE_STYLE == 5
 static constexpr uint32_t   BASE_FRAMERATE = 60; //10
-static constexpr float      BASE_FRAMETIME = 16.6666666667; //100
-static constexpr float      BASE_1_FRAMETIME = 0.05999999999; //0.01f   // 1/BASE_FRAMETIME
-static constexpr float      BASE_FRAMETIME_1000 = 0.01666666666; //0.1f    // BASE_FRAMETIME/1000
+static constexpr double     BASE_FRAMETIME = 16.6666666667; //100
+static constexpr double     BASE_1_FRAMETIME = 0.05999999999; //0.01f   // 1/BASE_FRAMETIME
+static constexpr double     BASE_FRAMETIME_1000 = 0.01666666666; //0.1f    // BASE_FRAMETIME/1000
 #endif
 
 //-----------------
